Add bcc and bpftrace tools
`bcc` and `bpftrace` provide many tools that are useful for obtaining
insights into various aspects of the system. Include them in the
container.

All the required tools are being built for ppc64le, so switch to using
the upstream packages in the `edge` repository. After the next
alpinelinux release, we can switch over to using alpine:latest

The image size has drastically increased, largely due to the
dependencies pulled in by bcc/bpftrace:
- python3 is ~45M
- clang is ~33M
- llvm is ~80M
- bcc itself is ~50M

A static build for bpftrace should help, but has many problems right now
and is being discussed in the upstream community:
- https://github.com/iovisor/bpftrace/issues/342

Signed-off-by: Naveen N. Rao <naveen.n.rao@linux.vnet.ibm.com>